+++ lightly_studio_view/src/lib/components/AnnotationDetails/AnnotationDetailsPanel/AnnotationDetailsPanel.svelte
@@ -37,7 +37,7 @@
             <AnnotationMetadata {annotationId} {onUpdate} />
 
             {#if sample}
-                <SampleMetadata {sample} showCustomMetadata={false} />
+                <SampleMetadata {sample} showCustomMetadata={false} showTags={true} />
 
                 <Button
                     variant="secondary"